gol/requirements.txt
numpy>=1.24,<3
Pillow>=10.0
streamlit>=1.28
streamlit-image-coordinates>=0.4.0
pytest>=7.0
pytest-cov>=4.0
